requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.3
Werkzeug>=3.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
qrcode[pil]>=7.4
Pillow>=10.0.0
gunicorn>=21.2.0
